fix(client): 🐛 Fix retreiving user's role
Fixed an issue with retreiving the player's role from the private player collection
