--- client/src/routes/winloss/+page.svelte
@@ -1,159 +1,0 @@
-<script lang="ts">
-  import { getAuth } from "firebase/auth";
-  import { doc, getDoc } from "firebase/firestore";
-  import { lobbyCollection, privPlayerCollection } from "$lib/firebase/firestore-collections";
-  import { page } from "$app/stores";
-
-  const auth = getAuth();
-  const user = auth.currentUser;
-  const currentUid:string = user!.uid;
-  let code: string = $page.url.search.split("=")[1];
-  
-  let end:string = "";
-  let aliveCount = 0;
-  let aliveCatCount = 0;
-  let displayname = "";
-
-  async function checkLobby() {
-    const lobby = doc(lobbyCollection, code);
-    const validLobby = await getDoc(lobby);
-
-    if (!validLobby.exists()) {
-      throw new Error("Invalid Lobby");
-    }
-
-  const { uids, players } = validLobby.data();
-
-  displayname = players[uids.indexOf(currentUid)].displayName;
-
-  // function to count the number of players currently alive
-    for(let i = 0; i < uids.length; i++) {
-      if(players[i].alive == true) {
-        aliveCount++;
-        // count the number of remaining players who are alive
-        let privatePlayer = doc(privPlayerCollection, uids[i]);
-        let validPlayer = await getDoc(privatePlayer);
-
-        let { role } = validPlayer.data();
-        if(role == "CAT") {
-          aliveCatCount++;
-          // within the alive players, count the number that are cats
-        }
-      }
-    }
-  let aliveCatFishCount = aliveCount - aliveCatCount;
-  // any other player who is alive and not a cat is a catfish
-  
-  let currentPlayer = doc(privPlayerCollection, currentUid);
-  let validCurrent = await getDoc(currentPlayer);
-
-  let { role } = validCurrent.data();
-  // set conditions for the game to be declared a win or loss
-    if(role == "CAT") {
-      if(aliveCatFishCount == 0) {
-        end = "Cat Win";
-      }
-      else if(aliveCatCount == aliveCatFishCount) {
-        end = "Cat Lose";
-      }
-    }
-    else {
-      if(aliveCatFishCount == 0) {
-        end = "Catfish Lose";
-      }
-      else if(aliveCatCount == aliveCatFishCount) {
-        end = "Catfish Win";
-      }
-    }
-}
-</script>
-
-<div class="container">
-  <!-- Check the user's role and display the correct win or loss page accordingly -->
-  {#if end = "Cat Win"}
-  <div class="banner">
-    <h2>Hooray!</h2>
-    <h2>You have sniffed out the cat fish!</h2>
-  </div>
-
-    <div class="image"></div>
-
-    <div class="displayname">
-      <h2>{displayname} was sentenced to 9 lives in purrison!</h2>
-    </div>
-  {/if}
-  {#if end = "Cat Lose"}
-    <div class="banner">
-      <h2>You have cat to be kitten me!</h2>
-      <h2>The impawster was not caught!</h2>
-   </div>
-
-    <div class="image"></div>
-
-    <div class="displayname">
-      <h2>{displayname} has taken over the litter!</h2>
-    </div>
-  {/if}
-  {#if end = "Catfish Win"}
-    <div class="banner">
-      <h2>O-fish-ally the greatest!</h2>
-      <h2>Way to deceive your furry friends!</h2>
-   </div>
-
-    <div class="image"></div>
-
-    <div class="displayname">
-      <h2>You have successfully taken over the litter!</h2>
-    </div> 
- 
-  {/if}
-  {#if end = "Catfish Lose"}
-    <div class="banner">
-      <h2>You have been caught (and released)!</h2>
-      <h2>Might as well have been a <span>clown</span> fish...</h2>
-   </div>
-
-    <div class="image"></div>
-
-    <div class="displayname">
-      <h2>You should go back to your sea ani-anim-aneme... home</h2>
-    </div> 
-  {/if}
-</div>
-
-<style>
-  .container {
-    position: relative;
-    height: 100vh;
-  }
-
-  .banner {
-    position: absolute;
-    width: 60%;
-    top: 10%;
-    left: 20%;
-    text-align: center;
-    font-size: 2.1em;
-  }
-
-  .image {
-    position: absolute;
-    width: 20%;
-    height: 30%;
-    top: 35%;
-    left: 40%;
-    border: 2px solid black;
-  }
-
-  .displayname {
-    position: absolute;
-    width: 80%;
-    top: 75%;
-    left: 15%;
-    font-size: 2.1em;
-  }
-
-  span {
-    font-weight: bold;
-  }
-</style>
